Creation de poll_answerController

+++ src/main/java/fr/utbm/fisa/communicationapi/infrastructure/repositories/Poll_answerRepository.java
@@ -3,5 +3,5 @@ package fr.utbm.fisa.communicationapi.infrastructure.repositories;
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Poll_answer;
 import org.springframework.data.repository.CrudRepository;
 
-public interface Poll_answerRepository extends CrudRepository<Poll_answer, Long> {
+public interface Poll_answerRepository extends CrudRepository<Poll_answer, Integer> {
 }
